Improve code sharing in cli module
