+++ tests/cli/test_api_gen.py
@@ -1,4 +1,3 @@
-import logging
 import os
 from datetime import datetime
 from pathlib import Path
@@ -6,10 +5,8 @@ from tempfile import TemporaryDirectory
 from unittest import mock
 
 import pytest
-import typer
 
 from openapi_spec_tools.cli.api_gen import generate_api
-from openapi_spec_tools.cli.api_gen import open_oas_with_error_handling
 from tests.helpers import StringIo
 from tests.helpers import asset_filename
 
@@ -26,27 +23,6 @@ def temp_working_dir():
 def _read_text(filename: str) -> str:
     with open(filename, "r", encoding="utf-8", newline="\n") as fp:
         return fp.read()
-
-
-@pytest.mark.parametrize(
-    ["filename", "message"],
-    [
-        pytest.param("gone", "ERROR: failed to find", id="missing"),
-        pytest.param("bad.json", "ERROR: unable to parse", id="bad-json"),
-        pytest.param("bad.yaml", "ERROR: unable to parse", id="bad-yaml"),
-    ]
-)
-def test_open_oas(filename, message) -> None:
-    logger = logging.getLogger("")
-    with (
-        mock.patch('sys.stdout', new_callable=StringIo) as mock_stdout,
-        pytest.raises(typer.Exit) as err,
-    ):
-        open_oas_with_error_handling(asset_filename(filename), logger)
-
-    assert err.value.exit_code == 1
-    output = mock_stdout.getvalue()
-    assert output.startswith(message)
 
 
 @pytest.mark.parametrize(
